+++ ksqldb-rest-app/src/main/java/io/confluent/ksql/rest/server/DefaultKsqlSecurityContextProvider.java
@@ -13,7 +13,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package io.confluent.ksql.api.endpoints;
+package io.confluent.ksql.rest.server;
 
 import io.confluent.kafka.schemaregistry.client.SchemaRegistryClient;
 import io.confluent.ksql.api.auth.ApiSecurityContext;
